Introduced URLRequestExtension to encapsulate the generation of a URLRequest WebUITableViewCell: Streamlined with URLRequestExtension, after migration to WKWebView the migration to WKNavigationDelegate  needed to be handled Signed-off-by: Tim Müller-Seydlitz <timbms@gmail.com>
